% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_base_url.R
\name{api_base_url}
\alias{api_base_url}
\title{Retrieve the Base URL for the data.public.lu API}
\usage{
api_base_url()
}
\value{
A character string representing the base URL of the data.public.lu API.
}
\description{
This function returns the base URL used to access the data.public.lu open data API.
It provides a consistent and easily accessible way to construct API requests.
}
\examples{
# Obtain the base URL for the API.
base_url <- api_base_url()
print(base_url)

# Construct a full API request for a specific endpoint (example).
endpoint <- "datasets"
full_url <- paste0(api_base_url(), "/", endpoint)
print(full_url)
}
