% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapubliclu-package.R
\docType{package}
\name{datapubliclu-package}
\alias{datapubliclu}
\alias{datapubliclu-package}
\title{datapubliclu: R Package for Accessing data.public.lu Datasets}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides a convenient interface for accessing datasets from https://data.public.lu, the open data portal of Luxembourg. It simplifies the process of downloading and importing data directly into your R environment.
}
\author{
\strong{Maintainer}: Christos Avrilionis \email{avrilionis@gmail.com} (\href{https://orcid.org/0009-0005-5444-8067}{ORCID})

}
\keyword{internal}
